@@ -4,11 +4,12 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --config ./jest.confing.js",
     "dev": "cross-env NODE_ENV=development webpack --mode development",
     "build": "cross-env NODE_ENV=production webpack --mode production",
     "watch": "cross-env NODE_ENV=development webpack --mode development --watch",
-    "start": "cross-env NODE_ENV=development webpack-dev-server --mode development --open"
+    "start": "cross-env NODE_ENV=development webpack-dev-server --mode development --open",
+    "test:watch": "jest --watch"
   },
   "browserslist": "> 0.25%, not dead",
   "author": "Dmitriy",
@@ -16,6 +17,9 @@
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@babel/plugin-proposal-class-properties": "^7.10.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-modules-commonjs": "^7.12.1",
+    "@babel/plugin-transform-runtime": "^7.12.10",
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
@@ -23,6 +27,9 @@
     "@eslint-kit/eslint-config-prettier": "^2.0.0",
     "@eslint-kit/eslint-config-react": "^2.0.0",
     "@eslint-kit/eslint-config-typescript": "^3.1.2",
+    "@types/enzyme": "^3.10.8",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@types/jest": "^26.0.19",
     "@types/react": "^16.9.51",
     "@types/react-dom": "^16.9.8",
     "@types/webpack": "^4.41.22",
@@ -39,6 +46,7 @@
     "eslint-plugin-react-hooks": "^4.1.2",
     "file-loader": "^6.1.1",
     "html-webpack-plugin": "^4.5.0",
+    "jest": "^26.6.3",
     "less": "^3.12.2",
     "less-loader": "^7.0.2",
     "mini-css-extract-plugin": "^1.0.0",
@@ -47,8 +55,9 @@
     "sass-loader": "^10.0.3",
     "style-loader": "^2.0.0",
     "terser-webpack-plugin": "^4.2.3",
+    "ts-jest": "^26.4.4",
     "ts-loader": "^8.0.4",
-    "typescript": "^4.0.3",
+    "typescript": "^4.1.3",
     "webpack": "^5.0.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.7.1",
@@ -56,10 +65,14 @@
   },
   "dependencies": {
     "@babel/polyfill": "^7.11.5",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@testing-library/react": "^11.2.2",
     "@types/react-router": "^5.1.8",
     "@types/react-router-dom": "^5.1.6",
     "@types/styled-components": "^5.1.5",
     "@typescript-eslint/parser": "^4.4.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router": "^5.2.0",
