@@ -1,20 +1,21 @@
 import React, { useEffect, useState } from 'react'
 import { Observable } from 'rxjs'
 import { Cards } from '@Components/UI'
-import { createInformationBoard } from '@Components/lib'
+import { createInformationCards } from '@Components/lib'
+import { InformationBoardType } from '@src/Interfaces'
 
 export const App = (): React.ReactElement => {
-	const [informationBoard, setInformationBoard] = useState<Observable<
-		(string | number)[]
+	const [informationCards, setInformationCards] = useState<Observable<
+		InformationBoardType[]
 	> | null>(null)
 
 	useEffect(() => {
-		setInformationBoard(createInformationBoard())
+		setInformationCards(createInformationCards())
 	}, [])
 
 	return (
 		<div>
-			<Cards cards={informationBoard} />
+			<Cards cards={informationCards} />
 		</div>
 	)
 }
