@@ -2,5 +2,6 @@ export * from './getRandomNumber'
 export * from './createYardstick'
 export * from './checkNA'
 export * from './createYardsticksBoard'
-export * from './createInformationBoard'
+export * from './createInformationCards'
 export * from './yardsticks'
+export * from './getYardstickValue'
